Refactor: Переименование HOC в соответствии с выполняемой ответственностью
