+++ src/components/app/app.jsx
@@ -4,15 +4,15 @@ import GenreQuestionScreen from "../genre-question-screen/genre-question-screen.
 import PropTypes from "prop-types";
 import React, {PureComponent} from "react";
 import WelcomeScreen from "../welcome-screen/welcome-screen.jsx";
-import withAudioPlayer from "../../hocs/with-audio-player/with-audio-player.js";
+import withActivePlayer from "../../hocs/with-active-player/with-active-player.js";
 import {ActionCreator} from "../../reducer.js";
 import {connect} from "react-redux";
 import {GameType} from "../../const.js";
 import {Switch, Route, BrowserRouter} from "react-router-dom";
 
 
-const GenreQuestionScreenWrapped = withAudioPlayer(GenreQuestionScreen);
-const ArtistQuestionScreenWrapped = withAudioPlayer(ArtistQuestionScreen);
+const GenreQuestionScreenWrapped = withActivePlayer(GenreQuestionScreen);
+const ArtistQuestionScreenWrapped = withActivePlayer(ArtistQuestionScreen);
 
 
 class App extends PureComponent {
