+++ src/hocs/with-active-player/with-active-player.js
